creates folders for community, nutrition, user, and exercise and moves respective api route files to those folders according to the colors inthe database schema
