+++ server/api/exercise/exercise_difficulties/index.js
