+++ server/api/exercise/exercise_equipment/index.js
