+++ server/api/exercise/exercises/index.js
