+++ server/api/exercise/muscle_groups/index.js
