+++ server/api/index.js
@@ -1,26 +1,26 @@
-const activity_levels = require("./activity_levels");
-const authentication = require("./users/auth");
-const exercise_difficulties = require("./exercise_difficulties");
-const exercise_equipment = require("./exercise_equipment");
-const exercise_types = require("./exercise_types");
-const genders = require("./genders");
-const goals = require("./goals");
-const meal_types = require("./meal_types");
-const muscle_groups = require("./muscle_groups");
-const user_tiers = require("./user_tiers");
-const bodyparts = require("./bodyparts");
-const exercises = require("./exercises");
-const users = require("./users");
-const workouts = require("./workouts");
-const food_images = require("./food_images");
-const tutorial_videos = require("./tutorial_videos");
-const user_videos = require("./user_videos");
-const workouts_exercises = require("./workouts_exercises");
-const community_posts = require("./community_posts");
-const community_comments = require("./community_comments");
-const exercise_bodyparts = require("./exercise_bodyparts");
-const foods_meals_users = require("./foods_meals_users")
-const nutrition = require("./nutrition");
+const activity_levels = require("./user/activity_levels");
+const authentication = require("./user/users/auth");
+const exercise_difficulties = require("./exercise/exercise_difficulties");
+const exercise_equipment = require("./exercise/exercise_equipment");
+const exercise_types = require("./exercise/exercise_types");
+const genders = require("./user/genders");
+const goals = require("./user/goals");
+const meal_types = require("./nutrition/meal_types");
+const muscle_groups = require("./exercise/muscle_groups");
+const user_tiers = require("./user/user_tiers");
+const bodyparts = require("./exercise/bodyparts");
+const exercises = require("./exercise/exercises");
+const users = require("./user/users");
+const workouts = require("./exercise/workouts");
+const food_images = require("./nutrition/food_images");
+const tutorial_videos = require("./exercise/tutorial_videos");
+const user_videos = require("./exercise/user_videos");
+const workouts_exercises = require("./exercise/workouts_exercises");
+const community_posts = require("./community/community_posts");
+const community_comments = require("./community/community_comments");
+const exercise_bodyparts = require("./exercise/exercise_bodyparts");
+const foods_meals_users = require("./nutrition/foods_meals_users");
+const nutrition = require("./nutrition/nutrition");
 
 module.exports = {
   activity_levels,
